Audiomod-Windows/requirements.txt
openai-whisper
 torch
 better_profanity